Make the RawData Debug implementation less misleading.
It was printing the length of the slice rather than the contents.
This could lead some people to believe that the slice contents
are always the same even when they're changing.

Print the actual contents instead, but limit it to 20 elements,
and show an ellipsis and the total count if there's more than 20.